+++ users/src/sdk/sdkconfiguration.py
@@ -24,9 +24,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.0.1'
-    gen_version: str = '2.194.1'
-    user_agent: str = 'speakeasy-sdk/python 2.0.1 2.194.1 1.0.0 wingspan_users'
+    sdk_version: str = '2.0.2'
+    gen_version: str = '2.209.0'
+    user_agent: str = 'speakeasy-sdk/python 2.0.2 2.209.0 1.0.0 wingspan_users'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: