+++ users/RELEASES.md
@@ -40,3 +40,13 @@ Based on:
 - [python v2.0.1] users
 ### Releases
 - [PyPI v2.0.1] https://pypi.org/project/wingspan_users/2.0.1 - users
+
+## 2023-12-04 01:26:38
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.2] users
+### Releases
+- [PyPI v2.0.2] https://pypi.org/project/wingspan_users/2.0.2 - users